aggiunto classi DTO

+++ src/main/java/it/perigea/provaSpring/service/LibroService.java
@@ -1,11 +1,17 @@
 package it.perigea.provaSpring.service;
 
+import java.util.List;
 import java.util.Optional;
+
+import javax.persistence.EntityNotFoundException;
+import javax.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import it.perigea.provaSpring.entities.Libro;
+import it.perigea.provaSpring.entities.dto.LibroDto;
+import it.perigea.provaSpring.entities.dto.LibroMapper;
 import it.perigea.provaSpring.repository.LibroRepository;
 
 @Service
@@ -13,34 +19,69 @@ public class LibroService {
 	
 	@Autowired
 	private LibroRepository repository;
+	
+	@Autowired
+	private LibroMapper mapper;
 
 	
-	public Iterable<Libro> getAllLibri() {
-		return  repository.findAll() ;
+	public List<LibroDto> getAllLibri() {
+		List<Libro> libri= repository.findAll();
+		//List<LibroDto> libriDto= new ArrayList<>();
+		//libri.forEach(libro -> libriDto.add(mapper.libroToLibroDto(libro)) );
+		List<LibroDto> libriDto=mapper.listaLibriToListaLibriDto(libri);
+		return   libriDto;
 	}
 	
-	public Optional<Libro> getById(Long id) {
-		return repository.findById(id);
-	}
+	public LibroDto getByIsbn(String isbn) {
+		Libro libro = repository.findByIsbn(isbn).orElseThrow( () -> new EntityNotFoundException("libro con isbn=" + isbn +" non presente nel DataBase."));
+		//Ho un oggetto Optional e se null solleva l'eccezione tramite la lamba function altrimenti restituisce l'oggetto contenuto (Libro).
+		LibroDto libroDto= mapper.libroToLibroDto( libro );
+		return libroDto;
+		}
 	
-	public Libro saveLibro(Libro libroDaSalvare) {
+	public LibroDto saveOrUpdateLibro(LibroDto libroDaSalvareDto) {
+		Optional<Libro> libroUpdate = repository.findByIsbn(libroDaSalvareDto.getIsbn());
+		Libro libroDaSalvare=new Libro();
+		if ( libroUpdate.isPresent() ){
+			libroDaSalvare = mapper.partialUpdate(libroDaSalvareDto, libroUpdate.get() );
+		} else {
+			libroDaSalvare = mapper.libroDtoToLibro(libroDaSalvareDto);
+		}
 		repository.save(libroDaSalvare);
-		return libroDaSalvare;
+		return mapper.libroToLibroDto(libroDaSalvare);	//posso restituire il libro preso in ingresso ma voglio vedere come va la doppia conversione.
 	}
 
-	public Libro deleteById(Long id) {
-		Libro libroDaEliminare=repository.findById(id).get();
+	@Transactional
+	public LibroDto deleteByIsbn(String isbn) {
+		Optional<Libro> libroDaEliminareOptional=repository.findByIsbn(isbn);	//al posto di optional e if si può usare orElseThrow() come sopra.
+		if (!libroDaEliminareOptional.isPresent() )	
+			throw new EntityNotFoundException("libro da eliminare con isbn=" + isbn +" non presente nel DataBase");
+		repository.deleteByIsbn(isbn);
+		LibroDto libroEliminato=mapper.libroToLibroDto( libroDaEliminareOptional.get() );
+		return libroEliminato;
+	}
+	
+	@Transactional
+	public LibroDto deleteById(Long id) {
+		Optional<Libro> libroDaEliminareOptional=repository.findById(id);	//al posto di optional e if si può usare orElseThrow() come sopra.
+		if (!libroDaEliminareOptional.isPresent() )	
+			throw new EntityNotFoundException("libro da eliminare con id=" + id +" non presente nel DataBase");
 		repository.deleteById(id);
-		return libroDaEliminare;
+		LibroDto libroEliminato=mapper.libroToLibroDto( libroDaEliminareOptional.get() );
+		return libroEliminato;
 	}
 	
-	public Libro deleteLibro(Libro libroDaEliminare) {
+	public LibroDto deleteLibro(LibroDto libroDaEliminareDto) {
+		Libro libroDaEliminare=mapper.libroDtoToLibro(libroDaEliminareDto);
+		libroDaEliminare.setId(repository.findByIsbn(libroDaEliminareDto.getIsbn()).get().getId());	//potevo direttamente farmi restituire il libro by isbn ed eliminarlo
 		repository.delete(libroDaEliminare);
-		return libroDaEliminare;
+		return mapper.libroToLibroDto(libroDaEliminare);
 	}
 	
-	public Iterable<Libro> getByTitolo(String titolo) {
-		return repository.findAllByTitolo(titolo);
+	public List<LibroDto> getByTitolo(String titolo) {
+		List<Libro> libri = repository.findAllByTitolo(titolo);
+		List<LibroDto> libriDto = mapper.listaLibriToListaLibriDto(libri);
+		return libriDto;
 	}
 	
 }
